+++ code/core/src/main/java/org/betonquest/betonquest/config/migrator/from3to4/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Migrations from 3.x to 4.x.
+ */
+package org.betonquest.betonquest.config.migrator.from3to4;